feat: upgrade rustlings to 5.4.0

+++ rust/rust/rustlings/exercises/functions/functions2.rs
@@ -5,7 +5,7 @@ fn main() {
     call_me(3);
 }
 
-fn call_me(num: i32) {
+fn call_me(num: usize) {
     for i in 0..num {
         println!("Ring! Call number {}", i + 1);
     }
